oilfield-ops-intelligence/requirements.txt
streamlit
pandas
numpy
plotly
folium
sqlalchemy
psycopg2-binary
requests
python-dotenv
apache-airflow
pytest
pydantic
loguru
streamlit-folium
great-expectations
